requirements.txt
fastapi~=0.96.0
protobuf
types-protobuf
googleapis-common-protos
googleapis-common-protos-stubs
protoc-gen-pyhttp
grpcio-tools
python-dotenv~=1.0.0
SQLAlchemy~=2.0.15
uvicorn~=0.22.0
PyJWT~=2.7.0
PyCryptoDome
starlette~=0.27.0
pydantic~=1.10.8
minio
pymysql